combine performance summary data
